Restructuring Added setup file to circumvent bad import issue
