+++ app/__init__.py
@@ -1,0 +1,6 @@
+from flask import Flask
+
+APP = Flask(__name__)
+APP.config['JSONIFY_PRETTYPRINT_REGULAR'] = False
+
+import app.api
