+++ app/api.py
@@ -1,7 +1,6 @@
 from flask import Flask, jsonify, request
-from .resources import resources
-APP = Flask(__name__)
-APP.config['JSONIFY_PRETTYPRINT_REGULAR'] = False
+from app import APP
+from app.resources import orders
 
 
 @APP.route('/orders', methods=['POST'])
@@ -14,12 +13,12 @@ def add_order():
     order_content = data.get('order_content')
     user = data.get('user')
     order_status = data.get('order_status')
-    if not any(i.get('order_id') == order_id for i in resources.orders):
-        resources.orders.append({'order_id': order_id,
-                                 'order_content': order_content,
-                                 'user': user,
-                                 'order_status': order_status})
-        return jsonify(resources.orders)
+    if not any(i.get('order_id') == order_id for i in orders):
+        orders.append({'order_id': order_id,
+                       'order_content': order_content,
+                       'user': user,
+                       'order_status': order_status})
+        return jsonify(orders)
     return 'Order already exists', 404
 
 if __name__ == '__main__':
