--- resources/resources.py
@@ -1,46 +1,0 @@
-orders = [{"order_id":60,
-			"order_content":"Little Chicken",
-			"user":"Eric",
-			"order_status":"Denied",
-			},
-			{"order_id":204863,
-			"order_content":"Big Chicken",
-			"user":"Mwachala",
-			"order_status":"Accepted",
-			},
-
-			
-
-			{'order_id':9989,
-			'order_content':'Medium Chicken',
-			'user':'Mutie',
-			'order_status':'Accepted',
-			},
-
-			{'order_id':485,
-			'order_content':'Smol Chicken',
-			'user':'Frank',
-			'order_status':'Denied',
-			},
-
-			{'order_id':87456,
-			'order_content':'Not a Chicken',
-			'user':'Jorja',
-			'order_status':'Accepted',
-			},
-
-			{'order_id':12,
-			'order_content':'Turkey',
-			'user':'Michelle',
-			'order_status':'Completed',
-			},
-
-			{'order_id':220020,
-			'order_content':'Bigger Turkey',
-			'user':'Dessa',
-			'order_status':'Completed',
-			}
-
-
-
-]
